Partially Solves #6
This enhances the search of employee, supervisor, community, and project for the "Add" new order window by allowing the user to filter through all the possible options through incomplete typing just as a search bar would. Implementing this feature on the parameters fon the orders main search bar remains pending.